+++ packages/terra-functional-testing/tests/jest/terra-cli/wdio/test-runner.test.js
@@ -52,6 +52,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         theme: 'terra-default-theme',
         updateScreenshots: true,
+        useSeleniumStandalonService: true,
       };
 
       await TestRunner.run(options);
@@ -176,6 +177,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         themes: ['terra-default-theme'],
         updateScreenshots: true,
+        useSeleniumStandalonService: true,
       });
 
       expect(TestRunner.run).toHaveBeenCalledWith({
@@ -194,6 +196,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         theme: 'terra-default-theme',
         updateScreenshots: true,
+        useSeleniumStandalonService: true,
       });
     });
   });